Remove save_post hooks to fix incompatibility issue Move activator method contents

+++ includes/class-point-tracker-deactivator.php
@@ -32,6 +32,10 @@ class Point_Tracker_Deactivator
     {
         global $wpdb;
 
+        if(!remove_all_actions('save_post')) {
+            wp_die("Could not remove save_post actions");
+        }
+
         delete_option('pt-require-login');
 
         $the_page = get_page_by_title("Challenge");
